:sparkles: Add New Version in Repository

+++ src/RuleCollection.php
@@ -7,7 +7,7 @@
  *
  * http://glpi-project.org
  *
- * @copyright 2015-2022 Teclib' and contributors.
+ * @copyright 2015-2023 Teclib' and contributors.
  * @copyright 2003-2014 by the INDEPNET Development Team.
  * @licence   https://www.gnu.org/licenses/gpl-3.0.html
  *
@@ -467,6 +467,8 @@ class RuleCollection extends CommonDBTM
         $p['condition'] = 0;
         $p['_glpi_tab'] = $options['_glpi_tab'];
         $rand           = mt_rand();
+        $p['display_criterias'] = false;
+        $p['display_actions']   = false;
 
         foreach (['inherited','childrens', 'condition'] as $param) {
             if (
@@ -477,9 +479,19 @@ class RuleCollection extends CommonDBTM
             }
         }
 
-        $rule             = $this->getRuleClass();
-        $display_entities = ($this->isRuleRecursive()
-                           && ($p['inherited'] || $p['childrens']));
+        foreach (['display_criterias', 'display_actions'] as $param) {
+            if (
+                isset($options[$param])
+            ) {
+                $p[$param] = $options[$param];
+            }
+        }
+
+        $rule              = $this->getRuleClass();
+        $display_entities  = ($this->isRuleRecursive()
+                            && ($p['inherited'] || $p['childrens']));
+        $display_criterias = $p['display_criterias'];
+        $display_actions   = $p['display_actions'];
 
        // Do not know what it is ?
         $canedit    = (self::canUpdate()
@@ -540,6 +552,12 @@ class RuleCollection extends CommonDBTM
         if ($use_conditions) {
             $colspan++;
         }
+        if ($display_criterias) {
+            $colspan++;
+        }
+        if ($display_actions) {
+            $colspan++;
+        }
 
         $can_sort = $canedit && $nb;
         if (count($this->RuleList->list)) {
@@ -564,8 +582,13 @@ class RuleCollection extends CommonDBTM
         if ($use_conditions) {
             $header_row .= "<th>" . __('Use rule for') . "</th>";
         }
+        if ($display_criterias) {
+            $header_row .= "<th>" . RuleCriteria::getTypeName(2) . "</th>";
+        }
+        if ($display_actions) {
+            $header_row .= "<th>" . RuleAction::getTypeName(2) . "</th>";
+        }
         $header_row .= "<th>" . __('Active') . "</th>";
-
         if ($display_entities) {
             $header_row .= "<th>" . Entity::getTypeName(1) . "</th>";
         }
@@ -577,7 +600,15 @@ class RuleCollection extends CommonDBTM
 
         echo "<tbody class='sortable-rules'>";
         for ($i = $p['start'],$j = 0; isset($this->RuleList->list[$j]); $i++,$j++) {
-            $this->RuleList->list[$j]->showMinimalForm($target, $i == 0, $i == $nb - 1, $display_entities, $p['condition']);
+            $this->RuleList->list[$j]->showMinimalForm(
+                $target,
+                $i == 0,
+                $i == $nb - 1,
+                $display_entities,
+                $p['condition'],
+                $display_criterias,
+                $display_actions
+            );
             Session::addToNavigateListItems($ruletype, $this->RuleList->list[$j]->fields['id']);
         }
         echo "</tbody>";
@@ -638,12 +669,13 @@ JAVASCRIPT;
             $url = $CFG_GLPI["root_doc"];
         }
 
-       //if rules provides an initRules method, then we're able to reset them
-        if (method_exists($this->getRuleClass(), 'initRules')) {
-            echo "<a class='btn btn-primary' id='reset_rules' href='" . $rule->getSearchURL() . "?reinit=true' " .
-            //does not work.
-            //"onClick='if(confirm(\"" . __s('All rules will be erased and recreated from scratch. Are you sure?')."\")) { return true } else { return false; };' " .
-            "title='" . __s("Remove all equipment import rules and recreate from defaults") . "'" .
+        // if rules provides has default rules, then we're able to reset them
+        $ruleclass = $this->getRuleClass();
+        if ($ruleclass instanceof Rule && $ruleclass->hasDefaultRules()) {
+            echo "<a class='btn btn-primary' id='reset_rules' href='" . $rule->getSearchURL() . "?reinit=true&subtype=" . $ruleclass->getType() . "' " .
+            "onClick='if(confirm(\"" . __s('Rules will be erased and recreated from default. Are you sure?') . "\"))
+            { return true } else { return false; };' " .
+            "title='" . __s("Delete all rules and recreate them by default") . "'" .
             ">" . __('Reset rules') . "</a>&nbsp;";
         }
         echo "<a class='btn btn-primary' href='#' data-bs-toggle='modal' data-bs-target='#allruletest$rand'>" .
@@ -651,7 +683,7 @@ JAVASCRIPT;
         Ajax::createIframeModalWindow(
             'allruletest' . $rand,
             $url . "/front/rulesengine.test.php?" .
-                                          "sub_type=" . $this->getRuleClassName() .
+                                          "sub_type=" . $ruleclass->getType() .
                                           "&condition=" . $p['condition'],
             ['title' => __('Test rules engine')]
         );
@@ -1167,6 +1199,7 @@ JAVASCRIPT;
             $tmprule = new $rule['sub_type']();
            //check entities
             if ($tmprule->isEntityAssign()) {
+                $rule['entities_id'] = $DB->escape(Html::entity_decode_deep($rule['entities_id']));
                 $entities_found = $entity->find(['completename' => $rule['entities_id']]);
                 if (empty($entities_found)) {
                     $rules_refused[$k_rule]['entity'] = true;
@@ -1419,6 +1452,7 @@ JAVASCRIPT;
      **/
     public static function processImportRules()
     {
+        global $DB;
         $ruleCriteria = new RuleCriteria();
         $ruleAction   = new RuleAction();
         $entity       = new Entity();
@@ -1462,6 +1496,7 @@ JAVASCRIPT;
             if (!$item->isEntityAssign()) {
                 $params['entities_id'] = 0;
             } else {
+                $rule['entities_id'] = $DB->escape($rule['entities_id']);
                 $entities_found = $entity->find(['completename' => $rule['entities_id']]);
                 if (!empty($entities_found)) {
                     $entity_found          = array_shift($entities_found);
@@ -1701,11 +1736,8 @@ JAVASCRIPT;
                     $output["result"][$rule->fields["id"]]["id"] = $rule->fields["id"];
                     $rule->process($input, $output, $params);
 
-                    if (
-                        $output["_rule_process"]
-                        && $this->stop_on_first_match
-                    ) {
-                        unset($output["_rule_process"]);
+                    if ((isset($output['_stop_rules_processing']) && (int) $output['_stop_rules_processing'] === 1) || ($output["_rule_process"] && $this->stop_on_first_match)) {
+                        unset($output["_stop_rules_processing"], $output["_rule_process"]);
                         $output["result"][$rule->fields["id"]]["result"] = 1;
                         $output["_ruleid"]                               = $rule->fields["id"];
                         return $output;
@@ -2353,6 +2385,10 @@ JAVASCRIPT;
                         'label'  => OperatingSystemArchitecture::getTypeName(Session::getPluralNumber()),
                         'link'   => 'ruledictionnaryoperatingsystemarchitecture.php',
                         'icon'   => OperatingSystemArchitecture::getIcon(),
+                    ], [
+                        'label'  => OperatingSystemEdition::getTypeName(Session::getPluralNumber()),
+                        'link'   => 'ruledictionnaryoperatingsystemedition.php',
+                        'icon'   => OperatingSystemEdition::getIcon(),
                     ]
                 ]
             ];
